Changed wrong varialbe type in travelDay, it was int, and it is double.

+++ OregonTrail/src/byui/cit260/oregontrail/model/ResourceScene.java
@@ -15,7 +15,7 @@ import java.util.Objects;
 public class ResourceScene extends RegularSceneType implements Serializable {
     private int amount;
     private ArrayList<InventoryItem> inventoryItem = new ArrayList<InventoryItem>();
-    public ResourceScene(String description, int travelDay, String river, String mountain, String weather, String month) {
+    public ResourceScene(String description, double travelDay, String river, String mountain, String weather, String month) {
         super(description, travelDay, river, mountain, weather, month);
     }
